ChannelLabel refactor, now shows the guild icon if necessary
